refactor(fieldmap): extract CSV download logic into a utility
