+++ js/source/entries/fieldmap_react.tsx
@@ -15,6 +15,7 @@ import {
     pearsonSkewness 
 } from '../include/fieldmap/utils/functions';
 import { printFieldMap } from '../include/fieldmap/utils/print';
+import { downloadLayoutCSV } from '../include/fieldmap/utils/downloadCsv';
 import { FieldMapLegend } from '../include/fieldmap/components/FieldMapLegend';
 import { FieldMapTooltip } from '../include/fieldmap/components/FieldMapTooltip';
 import { PlotLayer } from '../include/fieldmap/components/PlotLayer';
@@ -65,7 +66,7 @@ const FieldMapContainerInner: React.FC<FieldMapContainerProps> = ({
     } = useBorder();
 
     const {
-        plotObject, setPlotObject,
+        setPlotObject,
         plotList,
         dimensions, setDimensions,
         bounds, renderBounds
@@ -901,7 +902,7 @@ const FieldMapContainerInner: React.FC<FieldMapContainerProps> = ({
     const downloadHeatmapImage = () => {
         const svgEl = document.getElementById('fieldmap_chart_svg');
         if (!svgEl) return;
-        
+
         const svgString = new XMLSerializer().serializeToString(svgEl);
         const svgBlob = new Blob([svgString], { type: 'image/svg+xml;charset=utf-8' });
         const blobURL = URL.createObjectURL(svgBlob);
@@ -916,7 +917,7 @@ const FieldMapContainerInner: React.FC<FieldMapContainerProps> = ({
                 context.fillStyle = '#ffffff';
                 context.fillRect(0, 0, canvas.width, canvas.height);
                 context.drawImage(image, 0, 0);
-                
+
                 const pngData = canvas.toDataURL('image/png');
                 const downloadLink = document.createElement('a');
                 downloadLink.download = `${selectedViewLabel || 'fieldmap'}_heatmap.png`;
@@ -928,75 +929,7 @@ const FieldMapContainerInner: React.FC<FieldMapContainerProps> = ({
     };
 
     const handleDownloadCSV = () => {
-        let cols_csv_header = [];
-        for (let i = bounds.minCol; i <= bounds.maxCol; i++) {
-            cols_csv_header.push(i);
-        }
-        if (invertCols) {
-            cols_csv_header.reverse();
-        }
-        let csv = '';
-        csv += ['Rows/Columns', ...cols_csv_header].join(',') + '\n';
-
-        let coord_matrix: string[][] = [];
-        const sortedPlots = [...plotList].filter(p => p.type !== 'border');
-
-        sortedPlots.forEach(plot => {
-            const r = Number(plot.observationUnitPosition.positionCoordinateY) - bounds.minRow;
-            const c = Number(plot.observationUnitPosition.positionCoordinateX) - bounds.minCol;
-
-            if (!coord_matrix[r]) coord_matrix[r] = [];
-
-            let cellVal = '';
-            if (csvDownloadOpts.accession) {
-                cellVal += plot.germplasmName || plot.crossName || '';
-                if (plot.additionalInfo?.intercropGermplasm) {
-                    plot.additionalInfo.intercropGermplasm.forEach((g: any) => {
-                        cellVal += `, ${g.germplasmName}`;
-                    });
-                }
-            }
-            if (csvDownloadOpts.obsUnit && plot.observationUnitName) {
-                cellVal += (cellVal ? '\n' : '') + plot.observationUnitName;
-            }
-            if (csvDownloadOpts.plotId && plot.observationUnitDbId) {
-                cellVal += (cellVal ? '\n' : '') + plot.observationUnitDbId;
-            }
-            if (csvDownloadOpts.plotNum && plot.observationUnitPosition.observationLevel?.levelCode) {
-                cellVal += (cellVal ? '\n' : '') + plot.observationUnitPosition.observationLevel.levelCode;
-            }
-            if (csvDownloadOpts.familyName && plot.additionalInfo?.familyName) {
-                cellVal += (cellVal ? '\n' : '') + plot.additionalInfo?.familyName;
-            }
-            if (csvDownloadOpts.crossName && plot.crossName) {
-                cellVal += (cellVal ? '\n' : '') + plot.crossName;
-            }
-
-            coord_matrix[r][c] = `"${cellVal}"`;
-        });
-
-        if (!invertRows) {
-            coord_matrix.reverse();
-        }
-
-        coord_matrix.forEach((rowArr, idx) => {
-            if (!rowArr) rowArr = Array(bounds.numCols).fill('""');
-            for (let i = 0; i < bounds.numCols; i++) {
-                if (rowArr[i] === undefined) rowArr[i] = '""';
-            }
-            if (invertCols) {
-                rowArr.reverse();
-            }
-
-            const rowLabel = invertRows ? bounds.minRow + idx : bounds.maxRow - idx;
-            csv += [rowLabel, ...rowArr].join(',') + '\n';
-        });
-
-        const hiddenElement = document.createElement('a');
-        hiddenElement.href = 'data:text/csv;charset=utf-8,' + encodeURI(csv);
-        hiddenElement.target = '_blank';
-        hiddenElement.download = `Trial_${trialId}_spatial_layout.csv`;
-        hiddenElement.click();
+        downloadLayoutCSV(trialId, bounds, plotList, invertCols, invertRows, csvDownloadOpts);
         setShowDownloadCSVModal(false);
     };
